requirements.txt
django==1.11
pillow==4.3
django-ckeditor==5.4.0
sorl-thumbnail==12.4.1
django-hitcount==1.3.0
django-taggit==0.22.2
django-select2==6.0.2

